fix(guard): bind signed trust envelopes
