+++ src/codex_plugin_scanner/guard/inventory_contract.py
@@ -575,7 +575,9 @@ def _item_from_artifact(
         }
     )
     content_hash = _resolve_item_content_hash(safe_metadata, semantic_text)
-    safe_metadata = _apply_trust_attestation_metadata(
+    from .runtime.trust_attestation import apply_trust_attestation_metadata
+
+    safe_metadata = apply_trust_attestation_metadata(
         safe_metadata,
         agent_id=f"{harness}:local",
         item_id=artifact_id,
@@ -690,98 +692,6 @@ def _mcp_tool_items_from_artifact(
 
 def _string_value(value: object) -> str | None:
     return value if isinstance(value, str) else None
-
-
-def _apply_trust_attestation_metadata(
-    metadata: dict[str, object],
-    *,
-    agent_id: str,
-    item_id: str,
-    item_kind: InventoryItemKind,
-    content_hash: str,
-) -> dict[str, object]:
-    from .runtime.trust_attestation import (
-        build_trust_attestation_payload,
-        resolve_trust_attestation_signing_config,
-        sign_trust_attestation,
-    )
-
-    config = resolve_trust_attestation_signing_config()
-    if config is None:
-        return metadata
-
-    enriched = dict(metadata)
-
-    raw_trust_resolution = enriched.get("trustResolution")
-    if isinstance(raw_trust_resolution, dict):
-        trust_resolution = dict(raw_trust_resolution)
-        raw_resolution_metadata = trust_resolution.get("metadata")
-        resolution_metadata = dict(raw_resolution_metadata) if isinstance(raw_resolution_metadata, dict) else {}
-        evidence_hash = resolution_metadata.get("evidenceHash")
-        captured_at = trust_resolution.get("capturedAt")
-        if isinstance(evidence_hash, str) and evidence_hash and isinstance(captured_at, str) and captured_at:
-            resolution_metadata["attestation"] = sign_trust_attestation(
-                payload=build_trust_attestation_payload(
-                    agent_id=agent_id,
-                    item_id=item_id,
-                    item_kind=item_kind,
-                    content_hash=content_hash,
-                    captured_at=captured_at,
-                    evidence_hash=evidence_hash,
-                    scope="trust_resolution",
-                ),
-                config=config,
-                signed_at=captured_at,
-            )
-            resolution_metadata["attestationStatus"] = "signed"
-            trust_resolution["metadata"] = resolution_metadata
-            enriched["trustResolution"] = trust_resolution
-
-    raw_trust_layers = enriched.get("trustLayers")
-    if isinstance(raw_trust_layers, list):
-        signed_layers: list[object] = []
-        for raw_layer in raw_trust_layers:
-            if not isinstance(raw_layer, dict):
-                signed_layers.append(raw_layer)
-                continue
-            layer = dict(raw_layer)
-            raw_layer_metadata = layer.get("metadata")
-            layer_metadata = dict(raw_layer_metadata) if isinstance(raw_layer_metadata, dict) else {}
-            evidence_hash = layer_metadata.get("evidenceHash")
-            captured_at = layer.get("capturedAt")
-            layer_id = layer.get("layerId")
-            layer_type = layer.get("layerType")
-            if (
-                isinstance(evidence_hash, str)
-                and evidence_hash
-                and isinstance(captured_at, str)
-                and captured_at
-                and isinstance(layer_id, str)
-                and layer_id
-                and isinstance(layer_type, str)
-                and layer_type
-            ):
-                layer_metadata["attestation"] = sign_trust_attestation(
-                    payload=build_trust_attestation_payload(
-                        agent_id=agent_id,
-                        item_id=item_id,
-                        item_kind=item_kind,
-                        content_hash=content_hash,
-                        captured_at=captured_at,
-                        evidence_hash=evidence_hash,
-                        scope="trust_layer",
-                        layer_id=layer_id,
-                        layer_type=layer_type,
-                    ),
-                    config=config,
-                    signed_at=captured_at,
-                )
-                layer_metadata["attestationStatus"] = "signed"
-                layer["metadata"] = layer_metadata
-            signed_layers.append(layer)
-        enriched["trustLayers"] = signed_layers
-
-    return enriched
 
 
 def _first_present_value(mapping: dict[str, object], *keys: str) -> object | None:
